+++ lib/main.dart
@@ -7,12 +7,11 @@ import 'package:healingapp/pages/askQue.dart';
 // import 'package:healingapp/pages/signup.dart';
 // import 'package:healingapp/pages/signup2.dart';
 import 'package:healingapp/routes.dart';
-import 'package:healingapp/pages/login_page.dart';
 import 'package:healingapp/pages/home_page.dart';
 import 'package:healingapp/services/auth_service.dart';
 import 'package:healingapp/try.dart';
+import 'package:healingapp/userServices/SignUpPage.dart';
 import 'package:healingapp/userServices/loginPage.dart';
-import 'package:healingapp/userServices/signupPage.dart';
 import 'package:healingapp/userServices/verifyEmailPage.dart';
 import 'package:healingapp/widgets/bottomNavigatorBar.dart';
 import 'package:healingapp/try2.dart';
