--- lib/pages/login_page.dart
@@ -1,237 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-import 'package:lottie/lottie.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:healingapp/userServices/signupPage.dart';
-
-import '../widgets/bottomNavigatorBar.dart';
-
-void main() => runApp(LoginPage());
-
-
-
-String? existEmail, existPassword;
-final _formKey = GlobalKey<FormState>();
-
-class LoginPage extends StatefulWidget {
-  LoginPage({Key? key}) : super(key: key);
-  @override
-  State<LoginPage> createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  void click() {}
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SingleChildScrollView(
-        child: Container(
-          height: MediaQuery.of(context).size.height,
-          width: MediaQuery.of(context).size.width,
-          decoration: const BoxDecoration(
-              gradient: LinearGradient(
-                  begin: Alignment.topLeft,
-                  end: Alignment.bottomRight,
-                  colors: [
-                Colors.purpleAccent,
-                Colors.amber,
-                Colors.blue,
-              ])),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              const SizedBox(
-                height: 50,
-              ),
-              // SizedBox(
-              //   height: 200,
-              //   width: 300,
-              //   child: LottieBuilder.asset("assets/lottie/login2.json"),
-              // ),
-              const SizedBox(
-                height: 10,
-              ),
-              Container(
-                width: 325,
-                height: 470,
-                decoration: const BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.all(Radius.circular(15)),
-                ),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    const SizedBox(
-                      height: 30,
-                    ),
-                    const Text(
-                      "Healing App",
-                      style:
-                          TextStyle(fontSize: 28, fontWeight: FontWeight.bold),
-                    ),
-                    const SizedBox(
-                      height: 10,
-                    ),
-                    const Text(
-                      "Please Login",
-                      style: TextStyle(
-                        color: Colors.grey,
-                        fontSize: 15,
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 30,
-                    ),
-                    Container(
-                      width: 260,
-                      height: 60,
-                      key: _formKey,
-                      child: TextFormField(
-                        validator: (value) {
-                          if (value == null || value.isEmpty) {
-                            return 'Please enter some text';
-                          }
-                          return null;
-                        },
-                        decoration: InputDecoration(
-                            suffix: Icon(
-                              FontAwesomeIcons.envelope,
-                              color: Colors.red,
-                            ),
-                            labelText: "Email Address",
-                            border: OutlineInputBorder(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(8)),
-                            )),
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 12,
-                    ),
-                    Container(
-                      width: 260,
-                      height: 60,
-                      child: TextFormField(
-                        validator: (value) {
-                          if (value == null || value.isEmpty) {
-                            return 'Please enter some text';
-                          }
-                          return null;
-                        },
-                        obscureText: true,
-                        decoration: InputDecoration(
-                            suffix: Icon(
-                              FontAwesomeIcons.eyeSlash,
-                              color: Colors.red,
-                            ),
-                            labelText: "Password",
-                            border: OutlineInputBorder(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(8)),
-                            )),
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.fromLTRB(20, 0, 30, 0),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          TextButton(
-                            onPressed: click,
-                            child: const Text(
-                              "Forget Password",
-                              style: TextStyle(color: Colors.deepOrange),
-                            ),
-                          )
-                        ],
-                      ),
-                    ),
-                    GestureDetector(
-                      child: Container(
-                        alignment: Alignment.center,
-                        width: 250,
-                        decoration: const BoxDecoration(
-                            borderRadius: BorderRadius.all(Radius.circular(50)),
-                            gradient: LinearGradient(
-                                begin: Alignment.centerLeft,
-                                end: Alignment.centerRight,
-                                colors: [
-                                  Color(0xFF8A2387),
-                                  Color(0xFFE94057),
-                                  Color(0xFFF27121),
-                                ])),
-                        child: const Padding(
-                          padding: EdgeInsets.all(12.0),
-                          child: Text(
-                            'Login',
-                            style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 20,
-                                fontWeight: FontWeight.bold),
-                          ),
-                        ),
-                      ),
-                      onTap: () {
-                        // if (_formKey.currentState!.validate()) {
-                        // setState(() {
-                        //   existEmail = emailController.text;
-                        //   existPassword = passwordController.text;
-                        // });
-                        // signIn();
-                        // Navigator.pushReplacement(
-                        //     context,
-                        //     MaterialPageRoute(
-                        //       builder: (context) => bottomNavigatorBar(),
-                        //     ));
-                        // }
-                      },
-                    ),
-                    const SizedBox(
-                      height: 17,
-                    ),
-                    const Text(
-                      "Or Login using Social Media Account",
-                      style: TextStyle(
-                          fontWeight: FontWeight.bold, color: Colors.red),
-                    ),
-                    const SizedBox(
-                      height: 15,
-                    ),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                      children: [
-                        IconButton(
-                            onPressed: click,
-                            icon: const Icon(FontAwesomeIcons.facebook,
-                                color: Colors.blue)),
-                        IconButton(
-                            onPressed: click,
-                            icon: const Icon(
-                              FontAwesomeIcons.google,
-                              color: Colors.redAccent,
-                            )),
-                        IconButton(
-                            onPressed: click,
-                            icon: const Icon(
-                              FontAwesomeIcons.twitter,
-                              color: Colors.orangeAccent,
-                            )),
-                        IconButton(
-                            onPressed: click,
-                            icon: const Icon(
-                              FontAwesomeIcons.linkedinIn,
-                              color: Colors.green,
-                            ))
-                      ],
-                    )
-                  ],
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
