+++ lib/services/authHelper.dart
@@ -9,7 +9,6 @@ import 'package:google_sign_in/google_sign_in.dart';
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:healingapp/model/userModel.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:healingapp/userServices/signupPage.dart';
 
 class AuthHelper {
   static FirebaseAuth _auth = FirebaseAuth.instance;
