+++ lib/userServices/verifyEmailPage.dart
@@ -13,6 +13,7 @@ class VerifyEmailPage extends StatefulWidget {
 
 class _VerifyEmailPageState extends State<VerifyEmailPage> {
   bool isEmailVerified = false;
+  bool canResendEmail = false;
   Timer? timer;
 
   void initState() {
@@ -35,7 +36,7 @@ class _VerifyEmailPageState extends State<VerifyEmailPage> {
   }
 
   Future checkEmailVerified() async {
-    await FirebaseAuth.instance.currentUser!.reload();
+    // await FirebaseAuth.instance.currentUser!.reload();
     setState(() {
       isEmailVerified = FirebaseAuth.instance.currentUser!.emailVerified;
     });
@@ -48,6 +49,14 @@ class _VerifyEmailPageState extends State<VerifyEmailPage> {
     try {
       final user = FirebaseAuth.instance.currentUser;
       await user!.sendEmailVerification();
+
+      setState(
+        () => canResendEmail = false,
+      );
+      await Future.delayed(Duration(seconds: 5));
+      setState(
+        () => canResendEmail = true,
+      );
     } catch (e) {
       print(e);
       ScaffoldMessenger.of(context)
@@ -58,5 +67,48 @@ class _VerifyEmailPageState extends State<VerifyEmailPage> {
   @override
   Widget build(BuildContext context) => isEmailVerified
       ? bottomNavigatorBar()
-      : Scaffold(appBar: AppBar(title: Text("Verify Email")));
+      : Scaffold(
+          appBar: AppBar(title: Text("Verify Email")),
+          body: Padding(
+            padding: EdgeInsets.all(16),
+            child:
+                Column(mainAxisAlignment: MainAxisAlignment.center, children: [
+              Text(
+                "A Verification email has been sent to  your mail.",
+                style: TextStyle(fontSize: 20),
+                textAlign: TextAlign.center,
+              ),
+              SizedBox(
+                height: 24,
+              ),
+              ElevatedButton.icon(
+                onPressed: canResendEmail ? emailVerification : null,
+                style: ElevatedButton.styleFrom(
+                  minimumSize: Size.fromHeight(50),
+                ),
+                icon: Icon(
+                  Icons.email,
+                  size: 32,
+                ),
+                label: Text(
+                  "Resend Email",
+                  style: TextStyle(fontSize: 24),
+                ),
+              ),
+              SizedBox(
+                height: 24,
+              ),
+              TextButton(
+                onPressed: (() => FirebaseAuth.instance.signOut()),
+                style: ElevatedButton.styleFrom(
+                  minimumSize: Size.fromHeight(50),
+                ),
+                child: Text(
+                  "Cancel",
+                  style: TextStyle(fontSize: 24),
+                ),
+              )
+            ]),
+          ),
+        );
 }
