+++ lib/widgets/bottomNavigatorBar.dart
@@ -3,9 +3,7 @@ import 'package:healingapp/pages/askQue.dart';
 import 'package:healingapp/pages/helpUs.dart';
 import 'package:healingapp/pages/home_page.dart';
 import 'package:healingapp/userServices/loginPage.dart';
-import 'package:healingapp/pages/login_page.dart';
 import 'package:healingapp/pages/news_Page.dart';
-import 'package:healingapp/userServices/signupPage.dart';
 import 'package:healingapp/pages/yourQue.dart';
 import 'package:healingapp/try.dart';
 // import 'package:healingapp/pages/yourQue.dart';
