+++ lib/widgets/drawer.dart
@@ -1,69 +1,99 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 
-class MyDrawer extends StatelessWidget {
+class MyDrawer extends StatefulWidget {
   const MyDrawer({Key? key}) : super(key: key);
 
   @override
+  State<MyDrawer> createState() => _MyDrawerState();
+}
+
+class _MyDrawerState extends State<MyDrawer> {
+  final currentUser = FirebaseAuth.instance.currentUser;
+  // String userName = currentUser!.uid.toString();
+  
+
+  @override
   Widget build(BuildContext context) {
-    const userImageUrl = "https://raw.githubusercontent.com/RakeshBhardwaj-coder/RakeshBhardwaj.github.io/master/rks_logo.jpeg";
+    const userImageUrl =
+        "https://raw.githubusercontent.com/RakeshBhardwaj-coder/RakeshBhardwaj.github.io/master/rks_logo.jpeg";
     return Drawer(
       child: Container(
         color: Colors.blue,
         child: ListView(
+          // Important: Remove any padding from the ListView.
           padding: EdgeInsets.zero,
-          children:  const [
-            DrawerHeader(
+          children: [
+            const DrawerHeader(
               padding: EdgeInsets.zero,
               child: UserAccountsDrawerHeader(
-                
-
-                margin: EdgeInsets.zero,
-                  accountName: Text('Rakesh Bhardwaj'),
-                  accountEmail: Text("Rb610883@gmail.com"),
-                  currentAccountPicture:CircleAvatar( 
-                  backgroundImage:
-                      NetworkImage(userImageUrl),
-                ) ),
-
-
+                  margin: EdgeInsets.zero,
+                  accountName: Text("lund"),
+                  accountEmail: Text("Erak"),
+                  currentAccountPicture: CircleAvatar(
+                    backgroundColor: Colors.white,
+                    backgroundImage: NetworkImage(userImageUrl),
+                  )),
+            ),
+            Divider(
+              color: Colors.white,
             ),
             ListTile(
               leading: Icon(
                 CupertinoIcons.home,
                 color: Colors.white,
               ),
-              title: Text("Home",
-              style: TextStyle(
-                color: Colors.white,
+              title: Text(
+                "Home",
+                style: TextStyle(
+                  color: Colors.white,
+                ),
               ),
-              )
+              onTap: () {
+                // Update the state of the app.
+                // ...
+                // Then close the drawer
+                Navigator.pop(context);
+              },
             ),
             ListTile(
               leading: Icon(
                 CupertinoIcons.profile_circled,
                 color: Colors.white,
               ),
-              title: Text("Profile",
-              style: TextStyle(
-                color: Colors.white,
+              title: Text(
+                "Your Profile",
+                style: TextStyle(
+                  color: Colors.white,
+                ),
               ),
-              )
+              onTap: () {
+                // Update the state of the app.
+                // ...
+                Navigator.pop(context);
+              },
             ),
             ListTile(
               leading: Icon(
                 CupertinoIcons.mail,
                 color: Colors.white,
               ),
-              title: Text("Mail",
-              style: TextStyle(
-                color: Colors.white,
+              title: Text(
+                "SignOut",
+                style: TextStyle(
+                  color: Colors.white,
+                ),
               ),
-              )
+              onTap: () async {
+                // Update the state of the app.
+                // ...
+                await FirebaseAuth.instance.signOut();
+                Navigator.pop(context);
+              },
             ),
           ],
         ),
-        
       ),
     );
   }
